+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Circular symlinks no longer cause infinite loops when syncing a folder
+
 ## [1.21.0] - 2023-04-17
 
 ### Added